Create 01. Rectangle
Write a class for a rectangle object. It needs to have a width (Number), height (Number) and color (String) properties, which are set from the constructor and a calcArea() method, that calculates and returns the rectangle’s area.
Input
The constructor function will receive valid parameters.
Output
The calcArea() method should return a number. The first letter of color in output must be with upperCase.
Submit the class definition as is, without wrapping it in any function.